+++ wxc_sdk/as_api.py
@@ -72,10 +72,10 @@ __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'A
            'AsPrivateNetworkConnectApi', 'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi',
            'AsReceptionistContactsDirectoryApi', 'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRoomTabsApi',
            'AsRoomsApi', 'AsRouteGroupApi', 'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2UsersApi', 'AsScheduleApi',
-           'AsScimApiChild', 'AsScimV2Api', 'AsStatusAPI', 'AsTeamMembershipsApi', 'AsTeamsApi', 'AsTelephonyApi',
-           'AsTelephonyDevicesApi', 'AsTelephonyLocationApi', 'AsTransferNumbersApi', 'AsTrunkApi',
-           'AsVirtualLinesApi', 'AsVoiceMessagingApi', 'AsVoicePortalApi', 'AsVoicemailApi', 'AsVoicemailGroupsApi',
-           'AsVoicemailRulesApi', 'AsWebexSimpleApi', 'AsWebhookApi', 'AsWorkspaceDevicesApi',
+           'AsScimApiChild', 'AsScimV2Api', 'AsSequentialRingApi', 'AsStatusAPI', 'AsTeamMembershipsApi',
+           'AsTeamsApi', 'AsTelephonyApi', 'AsTelephonyDevicesApi', 'AsTelephonyLocationApi', 'AsTransferNumbersApi',
+           'AsTrunkApi', 'AsVirtualLinesApi', 'AsVoiceMessagingApi', 'AsVoicePortalApi', 'AsVoicemailApi',
+           'AsVoicemailGroupsApi', 'AsVoicemailRulesApi', 'AsWebexSimpleApi', 'AsWebhookApi', 'AsWorkspaceDevicesApi',
            'AsWorkspaceLocationApi', 'AsWorkspaceLocationFloorApi', 'AsWorkspaceNumbersApi',
            'AsWorkspacePersonalizationApi', 'AsWorkspaceSettingsApi', 'AsWorkspacesApi']
 
@@ -5896,6 +5896,7 @@ class AsPersonSettingsApiChild(AsApiChild, base=''):
             ('workspaces', 'outgoingPermission/digitPatterns'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'privacy'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'pushToTalk'): ('telephony/config/workspaces', '/'),
+            ('workspaces', 'sequentialRing'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'voicemail'): ('telephony/config/workspaces', '/'),
             ('people', 'agent'): ('telephony/config/people', '/'),
             ('people', 'callBridge'): ('telephony/config/people', '/features/'),
@@ -8938,15 +8939,14 @@ class AsScheduleApi(AsApiChild, base='telephony/config/locations'):
 
 class AsVoicemailApi(AsPersonSettingsApiChild):
     """
-    API for person's call voicemail settings. Also used for virtual lines
+    API for person's call voicemail settings. Also used for virtual lines and workspaces
     """
 
     feature = 'voicemail'
 
-    async def read(self, person_id: str, org_id: str = None) -> VoicemailSettings:
+    async def read(self, entity_id: str, org_id: str = None) -> VoicemailSettings:
         """
-        Read Voicemail Settings for a Person or virtual line
-        Retrieve a Person's Voicemail Settings
+        Read Voicemail Settings for an entity
 
         The voicemail feature transfers callers to voicemail based on your settings. You can then retrieve voice
         messages via Voicemail. Voicemail audio is sent in Waveform Audio File Format, .wav, format.
@@ -8957,22 +8957,21 @@ class AsVoicemailApi(AsPersonSettingsApiChild):
         This API requires a full, user, or read-only administrator auth token with a scope of spark-admin:people_read
         or a user auth token with spark:people_read scope can be used by a person to read their settings.
 
-        :param person_id: Unique identifier for the person or virtual line
-        :type person_id: str
-        :param org_id: Person is in this organization. Only admin users of another organization (such as partners)
+        :param entity_id: Unique identifier for the entity
+        :type entity_id: str
+        :param org_id: Entity is in this organization. Only admin users of another organization (such as partners)
             may use this parameter as the default is the same organization as the token used to access API.
         :type org_id: str
-        :return: user's voicemail settings
+        :return: entity's voicemail settings
         :rtype: VoicemailSettings
         """
-        url = self.f_ep(person_id=person_id)
+        url = self.f_ep(entity_id)
         params = org_id and {'orgId': org_id} or None
         return VoicemailSettings.model_validate(await self.get(url, params=params))
 
-    async def configure(self, person_id: str, settings: VoicemailSettings, org_id: str = None):
+    async def configure(self, entity_id: str, settings: VoicemailSettings, org_id: str = None):
         """
-        Configure Voicemail Settings for a Person
-        Configure a person's Voicemail Settings
+        Configure Voicemail Settings for an entity
 
         The voicemail feature transfers callers to voicemail based on your settings. You can then retrieve voice
         messages via Voicemail. Voicemail audio is sent in Waveform Audio File Format, .wav, format.
@@ -8985,23 +8984,19 @@ class AsVoicemailApi(AsPersonSettingsApiChild):
         :return:
         """
         # some settings can't be part of an update
-        data = settings.model_dump_json(exclude={'send_busy_calls': {'greeting_uploaded': True},
-                                                 'send_unanswered_calls': {'system_max_number_of_rings': True,
-                                                                           'greeting_uploaded': True},
-                                                 'voice_message_forwarding_enabled': True
-                                                 })
-        url = self.f_ep(person_id=person_id)
+        data = settings.update()
+        url = self.f_ep(entity_id)
         params = org_id and {'orgId': org_id} or None
-        await self.put(url, data=data, params=params)
+        await self.put(url, json=data, params=params)
 
-    async def _configure_greeting(self, *, person_id: str, content: Union[BufferedReader, str],
+    async def _configure_greeting(self, *, entity_id: str, content: Union[BufferedReader, str],
                             upload_as: str = None, org_id: str = None,
                             greeting_key: str):
         """
         handle greeting configuration
 
-        :param person_id: Unique identifier for the person.
-        :type person_id: str
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
         :param content: the file to be uploaded, can be a path to a file or a buffered reader (opened file); if a
             reader referring to an open file is passed then make sure to open the file as binary b/c otherwise the
             content length might be calculated wrong
@@ -9023,7 +9018,7 @@ class AsVoicemailApi(AsPersonSettingsApiChild):
             if not upload_as:
                 raise ValueError('upload_as is required')
         encoder = MultipartEncoder({'file': (upload_as, content, 'audio/wav')})
-        ep = self.f_ep(person_id=person_id, path=f'actions/{greeting_key}/invoke')
+        ep = self.f_ep(entity_id, path=f'actions/{greeting_key}/invoke')
         params = org_id and {'orgId': org_id} or None
         try:
             await self.post(ep, data=encoder, headers={'Content-Type': encoder.content_type},
@@ -9032,20 +9027,18 @@ class AsVoicemailApi(AsPersonSettingsApiChild):
             if must_close:
                 content.close()
 
-    def configure_busy_greeting(self, person_id: str, content: Union[BufferedReader, str],
+    def configure_busy_greeting(self, entity_id: str, content: Union[BufferedReader, str],
                                 upload_as: str = None, org_id: str = None):
         """
-        Configure Busy Voicemail Greeting for a Person
-        Configure a Person's Busy Voicemail Greeting by uploading a Waveform Audio File Format, .wav, encoded audio
-        file.
+        Configure Busy Voicemail Greeting for an entity
 
         Your request will need to be a multipart/form-data request rather than JSON, using the audio/wav Content-Type.
 
         This API requires a full or user administrator auth token with the spark-admin:people_write scope or a user
         auth token with spark:people_write scope can be used by a person to update their settings.
 
-        :param person_id: Unique identifier for the person.
-        :type person_id: str
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
         :param content: the file to be uploaded, can be a path to a file or a buffered reader (opened file); if a
             reader referring to an open file is passed then make sure to open the file as binary b/c otherwise the
             content length might be calculated wrong
@@ -9056,14 +9049,15 @@ class AsVoicemailApi(AsPersonSettingsApiChild):
             may use this parameter as the default is the same organization as the token used to access API.
         :type org_id: str
         """
-        self._configure_greeting(person_id=person_id, content=content, upload_as=upload_as, org_id=org_id,
+        self._configure_greeting(entity_id=entity_id, content=content, upload_as=upload_as, org_id=org_id,
                                  greeting_key='uploadBusyGreeting')
 
-    def configure_no_answer_greeting(self, person_id: str, content: Union[BufferedReader, str],
+    def configure_no_answer_greeting(self, entity_id: str, content: Union[BufferedReader, str],
                                      upload_as: str = None, org_id: str = None):
         """
-        Configure No Answer Voicemail Greeting for a Person
-        Configure a Person's No Answer Voicemail Greeting by uploading a Waveform Audio File Format, .wav, encoded
+        Configure No Answer Voicemail Greeting for an entity
+
+        Configure an entity's No Answer Voicemail Greeting by uploading a Waveform Audio File Format, .wav, encoded
         audio file.
 
         Your request will need to be a multipart/form-data request rather than JSON, using the audio/wav Content-Type.
@@ -9071,8 +9065,8 @@ class AsVoicemailApi(AsPersonSettingsApiChild):
         This API requires a full or user administrator auth token with the spark-admin:people_write scope or a user
         auth token with spark:people_write scope can be used by a person to update their settings.
 
-        :param person_id: Unique identifier for the person.
-        :type person_id: str
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
         :param content: the file to be uploaded, can be a path to a file or a buffered reader (opened file); if a
             reader referring to an open file is passed then make sure to open the file as binary b/c otherwise the
             content length might be calculated wrong
@@ -9083,18 +9077,18 @@ class AsVoicemailApi(AsPersonSettingsApiChild):
             may use this parameter as the default is the same organization as the token used to access API.
         :type org_id: str
         """
-        self._configure_greeting(person_id=person_id, content=content, upload_as=upload_as, org_id=org_id,
+        self._configure_greeting(entity_id=entity_id, content=content, upload_as=upload_as, org_id=org_id,
                                  greeting_key='uploadNoAnswerGreeting')
 
-    async def modify_passcode(self, person_id: str, passcode: str, org_id: str = None):
+    async def modify_passcode(self, entity_id: str, passcode: str, org_id: str = None):
         """
-        Modify a person's voicemail passcode.
+        Modify a entity's voicemail passcode.
 
-        Modifying a person's voicemail passcode requires a full administrator, user administrator or location
+        Modifying a entity's voicemail passcode requires a full administrator, user administrator or location
         administrator auth token with a scope of `spark-admin:telephony_config_write`.
 
-        :param person_id: Modify voicemail passcode for this person.
-        :type person_id: str
+        :param entity_id: Modify voicemail passcode for this person.
+        :type entity_id: str
         :param passcode: Voicemail access passcode. The minimum length of the passcode is 6 and the maximum length is
             30.
         :type passcode: str
@@ -9107,7 +9101,7 @@ class AsVoicemailApi(AsPersonSettingsApiChild):
             params['orgId'] = org_id
         body = dict()
         body['passcode'] = passcode
-        url = self.f_ep(person_id, 'passcode')
+        url = self.f_ep(entity_id, 'passcode')
         await super().put(url, params=params, json=body)
 
     async def reset_pin(self, person_id: str, org_id: str = None):
@@ -20149,6 +20143,205 @@ class AsAnonCallsApi(AsPersonSettingsApiChild):
         await super().put(url, params=params, json=body)
 
 
+class AsSequentialRingApi(AsPersonSettingsApiChild):
+    """
+    API for sequential ring settings
+
+    For now only used for workspaces
+    """
+
+    feature = 'sequentialRing'
+
+    async def read_criteria(self, entity_id: str, id: str,
+                      org_id: str = None) -> SequentialRingCriteria:
+        """
+        Retrieve sequential ring criteria for a workspace.
+
+        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
+        receives incoming calls, these numbers will ring one after another.
+        The sequential ring criteria specify settings such as schedule and incoming numbers for which to sequentially
+        ring or not.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
+        :param id: Unique identifier for the criteria.
+        :type id: str
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: :class:`SequentialRingCriteria`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.f_ep(entity_id, f'criteria/{id}')
+        data = await super().get(url, params=params)
+        r = SequentialRingCriteria.model_validate(data)
+        return r
+
+    async def configure_criteria(self, entity_id: str, id: str, settings: SequentialRingCriteria,
+                           org_id: str = None):
+        """
+        Modify sequential ring criteria for a workspace.
+
+        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
+        receives incoming calls, these numbers will ring one after another.
+        The sequential ring criteria specify settings such as schedule and incoming numbers for which to sequentially
+        ring or not.
+
+        This API requires a full, user or location administrator auth token with the `spark-admin:workspaces_write` to
+        update workspace settings.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param entity_id: Unique identifier for the workspace.
+        :type entity_id: str
+        :param id: Unique identifier for the criteria.
+        :type id: str
+        :param settings: new settings to be applied.
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = settings.update()
+
+        url = self.f_ep(entity_id, f'criteria/{id}')
+        await super().put(url, params=params, json=body)
+
+    async def delete_criteria(self, workspace_id: str, id: str, org_id: str = None):
+        """
+        Delete sequential ring criteria for a workspace.
+
+        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
+        receives incoming calls, these numbers will ring one after another.
+        The sequential ring criteria specify settings such as schedule and incoming numbers for which to sequentially
+        ring or not.
+
+        This API requires a full, read-only or location administrator auth token with a scope of
+        `spark-admin:workspaces_read` to read workspace settings.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param workspace_id: Unique identifier for the workspace.
+        :type workspace_id: str
+        :param id: Unique identifier for the criteria.
+        :type id: str
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.f_ep(workspace_id, f'criteria/{id}')
+        await super().delete(url, params=params)
+
+    async def create_criteria(self, workspace_id: str, settings: SequentialRingCriteria,
+                        org_id: str = None) -> str:
+        """
+        Create sequential ring criteria for a workspace.
+
+        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
+        receives incoming calls, these numbers will ring one after another.
+        The sequential ring criteria specify settings such as schedule and incoming numbers for which to sequentially
+        ring or not.
+
+        This API requires a full, user or location administrator auth token with the `spark-admin:workspaces_write` to
+        update workspace settings.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param workspace_id: Unique identifier for the workspace.
+        :type workspace_id: str
+        :param settings: new settings to be applied.
+        :type settings: SequentialRingCriteria
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: str
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = settings.update()
+        body.pop('id', None)
+
+        url = self.f_ep(workspace_id, f'criteria')
+        data = await super().post(url, params=params, json=body)
+        r = data['id']
+        return r
+
+    async def read(self, workspace_id: str,
+             org_id: str = None) -> SequentialRing:
+        """
+        Retrieve sequential ring settings for a workspace.
+
+        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
+        receives incoming calls, these numbers will ring one after another.
+
+        This API requires a full, read-only or location administrator auth token with a scope of
+        `spark-admin:workspaces_read` to read workspace settings.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param workspace_id: Unique identifier for the workspace.
+        :type workspace_id: str
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: :class:`SequentialRing`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.f_ep(workspace_id)
+        data = await super().get(url, params=params)
+        r = SequentialRing.model_validate(data)
+        return r
+
+    async def configure(self, workspace_id: str, settings: SequentialRing,
+                  org_id: str = None):
+        """
+        Modify sequential ring settings for a workspace.
+
+        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
+        receives incoming calls, these numbers will ring one after another.
+
+        This API requires a full, user or location administrator auth token with the `spark-admin:workspaces_write` to
+        update workspace settings.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param workspace_id: Unique identifier for the workspace.
+        :type workspace_id: str
+        :param settings
+        :type settings: SequentialRing
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = settings.update()
+        url = self.f_ep(workspace_id)
+        await super().put(url, params=params, json=body)
+
+
 class AsWorkspaceDevicesApi(AsApiChild, base='telephony/config/workspaces'):
     def list_gen(self, workspace_id: str, org_id: str = None) -> AsyncGenerator[TelephonyDevice, None, None]:
         """
@@ -20285,6 +20478,7 @@ class AsWorkspaceSettingsApi(AsApiChild, base='workspaces'):
     permissions_out: AsOutgoingPermissionsApi
     privacy: AsPrivacyApi
     push_to_talk: AsPushToTalkApi
+    sequential_ring: AsSequentialRingApi
     voicemail: AsVoicemailApi
 
     def __init__(self, session: AsRestSession):
@@ -20306,6 +20500,7 @@ class AsWorkspaceSettingsApi(AsApiChild, base='workspaces'):
         self.permissions_out = AsOutgoingPermissionsApi(session=session, selector=ApiSelector.workspace)
         self.privacy = AsPrivacyApi(session=session, selector=ApiSelector.workspace)
         self.push_to_talk = AsPushToTalkApi(session=session, selector=ApiSelector.workspace)
+        self.sequential_ring = AsSequentialRingApi(session=session, selector=ApiSelector.workspace)
         self.voicemail = AsVoicemailApi(session=session, selector=ApiSelector.workspace)