+++ docs/changes.rst
@@ -29,6 +29,7 @@ Release history
 - feat: do not disturb API for workspaces: :attr:`api.workspace_settings.dnd <wxc_sdk.workspace_settings.WorkspaceSettingsApi.dnd>`
 - feat: push to talk API for workspaces: :attr:`api.workspace_settings.push_to_talk <wxc_sdk.workspace_settings.WorkspaceSettingsApi.push_to_talk>`
 - feat: voicemail settings API for workspaces: :attr:`api.workspace_settings.voicemail <wxc_sdk.workspace_settings.WorkspaceSettingsApi.voicemail>`
+- feat: sequential ring settings API for workspaces: :attr:`api.workspace_settings.sequential_ring <wxc_sdk.workspace_settings.WorkspaceSettingsApi.sequential_ring>`
 
 1.19.0
 ------